refactor: move preferences to it's new route
